Fix timeout of binary Semaphore
The xSemaphoreTake() function was set to portTick_PERIOD_MS that would
cause it to timeout. This caused the interruptFunction to be triggered
without a hardware interrupt.
This change corrected that by replacing it with portMAX_DELAY which
stops the xSemaphoreTake() function from timing out and will now only
trigger on HW interrupt.